==> R_analysis/scr/00_libraries.R <==
renv::install("boot")
renv::install("cluster")
renv::install("codetools")
renv::install("lattice")
renv::install("MASS")
renv::install("Matrix")
renv::install("mgcv")
renv::install("nlme")
renv::install("nnet")
renv::install("survival")

renv::install("tidyverse")
renv::install("chris-mcginnis-ucsf/DoubletFinder")
renv::install("SoupX")
renv::install("cli")
renv::install("openssl")
renv::install("devtools")
renv::install("magick")
renv::install("ggbreak")
renv::install("lemon")
renv::install("bioc::ashr")
# renv::install("SeuratObject@4.1.4")
renv::install("SeuratObject")

# renv::install("bioc::msigdbr")
# renv::install("bioc::GenomeInfoDb")
# renv::install("bioc::ensembldb")
# renv::install("bioc::AnnotationHub")
renv::install("bioc::limma")
renv::install("bioc::scater")
renv::install("bioc::IRanges")
renv::install("bioc::glmGamPoi")
renv::install("bioc::miloR")
renv::install("bioc::multtest")

renv::install("ggdendro")
renv::install("bioc::ggtree")

renv::install("metap")

# devtools::install_version("crossmatch", version = "1.3.1", repos = "http://cran.us.r-project.org")
# devtools::install_version("multicross", version = "2.1.0", repos = "http://cran.us.r-project.org")

renv::install("jokergoo/ComplexHeatmap")
renv::install("satijalab/seurat-data")
renv::install("satijalab/azimuth")
# renv::install("jackbibby1/SCPA")

renv::install("bioc::scran")
renv::install("ggnewscale")

renv::install("bioc::DropletUtils")
renv::install("bioc::glmGamPoi")

renv::install("sf")
renv::install("harmony")
renv::install("satijalab/seurat-wrappers@d28512f804d5fe05e6d68900ca9221020d52cf1d")

renv::install("bioc::clusterProfiler")
renv::install("bioc::GSVA")
renv::install("bioc::msigdbr")

renv::install("homologene")
renv::install("lemon")
renv::install("multtest")
renv::install("metap")
renv::install("ggbreak")
renv::install("finalfit")
renv::install("enrichR")
renv::install("bioc::DESeq2")
renv::install("RNAseqQC")
renv::install("magick")
renv::install("UpSetR")

renv::install("devtools")
devtools::install_version("crossmatch", version = "1.3.1", repos = "http://cran.us.r-project.org")
devtools::install_version("multicross", version = "2.1.0", repos = "http://cran.us.r-project.org")
renv::install("igordot/clustermole@v1.1.0")
renv::install("curl")
renv::install("bioc::SummarizedExperiment@1.30")
renv::install("jackbibby1/SCPA")

renv::install("bioc::dittoSeq")

library(BiocManager)
