==> R_analysis/scr/00_soupX.R <==
# AIM ---------------------------------------------------------------------
# the aim of this script is to:
# 1. transfer the minimal number of files from the source directory to the project folder.
# 2. run SoupX and generate the cleaned raw matrices.

# libraries ---------------------------------------------------------------
library(tidyverse)
library(SoupX)
library(DropletUtils)

# transfer the count data to the project folder for simplicity ------------
# # define all the samples
# sample_id <- dir("/beegfs/scratch/ric.absinta/ric.absinta/analysis/BS_run04/cellranger7/output/") %>%
#   str_subset(pattern = "\\.",negate = T) %>%
#   str_subset(pattern = "multiqc_data|web_summaries",negate = T)
# 
# # crate the destination folders
# # crossing(sample_id = sample_id,
# #          folder = c("raw_feature_bc_matrix","filtered_feature_bc_matrix","analysis")) %>%
# #   pmap(function(sample_id,folder){
# #     dir.create(path = paste0("../../data/cellranger7_out/",sample_id,"/",folder),recursive = T)
# #     })
# sample_id %>%
#   map(function(sample_id){
#     dir.create(path = paste0("../../data/cellranger7_out/",sample_id),recursive = T)
#   })
# 
# # location of the original count matrices
# LUT_sample <- data.frame(source = "/beegfs/scratch/ric.absinta/ric.absinta/analysis/BS_run04/cellranger7/output/",
#                            destination = "/beegfs/scratch/ric.cosr/pedrini.edoardo/scRNAseq_Brainsphere_Absinta/BS_run_04/data/cellranger7_out",
#                            sample_id = sample_id) %>%
#   mutate(source_folder = paste0(source,sample_id,"/outs"))
# 
# # build the table to transfer the main folders
# LUT_transfer_folder <- LUT_sample %>%
#   group_by(source_folder) %>%
#   mutate(source_file = map(source_folder,function(x){
#     dir(x) %>% str_subset(pattern = "raw_feature_bc_matrix|filtered_feature_bc_matrix|analysis") %>%
#       str_subset(pattern = "\\.h5",negate = T)
#   })) %>%
#   unnest(source_file) %>%
#   mutate(source_file_full = paste0(source_folder,"/",source_file)) %>%
#   mutate(destination_file_full = paste0(destination,"/",sample_id,"/",source_file))
# 
# # copy the files from source to destination
# LUT_transfer_folder %>%
#   split(f = .$destination_file_full) %>%
#   lapply(function(x){
#     R.utils::copyDirectory(from = x$source_file_full,
#                            to = x$destination_file_full)
#   })
# 
# # copy and tranfer individual files
# dir("/beegfs/scratch/ric.absinta/ric.absinta/analysis/BS_run04/cellranger7/output/W8_CSF/outs")
# LUT_transfer_file <- LUT_sample %>%
#   group_by(source_folder) %>%
#   mutate(source_file = map(source_folder,function(x){
#     dir(x) %>% str_subset(pattern = "metrics_summary.csv") %>%
#       str_subset(pattern = "\\.h5",negate = T)
#   })) %>%
#   unnest(source_file) %>%
#   mutate(source_file_full = paste0(source_folder,"/",source_file)) %>%
#   mutate(destination_file_full = paste0(destination,"/",sample_id,"/",source_file))
# 
# file.copy(from = LUT_transfer_file$source_file_full,to = LUT_transfer_file$destination_file_full,overwrite = T)

# run the soupx processing ------------------------------------------------
id_sample <- dir("../../data/cellranger7_out/")

# do the preprocessing over all the dataset and save the output matrices
# x <- "W8_CSF"
# file <- paste0("../../raw_data/cellranger/",x,"/outs/")
# test <- load10X(dataDir = file)

# total
# "W8_CSF"
# "W8_CSF_alfa_lipoic_acid"
# "W8_CSF_dasatinib"
# "W8_CSF_ibudilast"
# "W8_CSF_tolebrutinib"
# "W8_IgG"
# "W8_IgG_tolebrutinib"
# "W8_low_cytokines"
# "W8_untreated"

# works
# "W8_CSF"
# "W8_CSF_alfa_lipoic_acid"
# "W8_CSF_dasatinib"
# "W8_CSF_ibudilast"
# "W8_IgG"
# "W8_IgG_tolebrutinib"
# "W8_low_cytokines"
# "W8_untreated"

# fail. Extremely high contamination estimated (0.63).
# "W8_CSF_tolebrutinib"

lapply(id_sample,function(x){
  # track the progress
  print(x)
  # define the location of the output of cellranger
  file <- paste0("../../data/cellranger7_out/",x)
  # Load data and estimate soup profile
  sc <- load10X(file)
  # Estimate rho
  sc <- autoEstCont(sc)
  # Clean the data
  out <- adjustCounts(sc,roundToInt = T)
  # save the data
  DropletUtils:::write10xCounts(paste0("../../data/SoupX_default_cellranger7/",x), out)
})

# test 01 failed sample ---------------------------------------------------
# I have followed the comment in the SoupX support page https://github.com/constantAmateur/SoupX/issues/60
# "If your data really does have 75% contamination, you probably shouldn't use it as that level of contamination likely indicates something went very wrong in the experiment. To check, I would look at the plot generated by autoEstCont. Does it have two peaks of roughly equal height, with one being around .75? If so, try setting your contamination to the location of the lower peak and proceeding with your analysis. The other thing I'd do is make extensive use of plotMarkerMap to see what the expression ratio to the soup looks like for a few genes that are commonly contamination. Without knowing your experiment it's hard to say what these are likely to be, but HB and IG genes usually work."

# based on the first plot generated from autoEstCont, I estimate a range of contamination between 0 and 0.2
# id_sample2 <- x <- "W8_CSF_tolebrutinib"
# 
# file <- paste0("../../data/cellranger7_out/",x)
# # Load data and estimate soup profile
# sc <- load10X(file)
# 
# # Estimate rho
# # fails
# sc2 <- autoEstCont(sc)
# # manual set rho
# sc3 <- setContaminationFraction(sc,contFrac = 0.1)
# # reduce the range for the estimate
# sc4 <- autoEstCont(sc,contaminationRange = c(0.01,0.5))
# 
# # plotMarkerMap(sc4,geneSet = "OLIG1") + plotMarkerMap(sc,geneSet = "OLIG1")
# out3 <- adjustCounts(sc3,roundToInt = T)
# out4 <- adjustCounts(sc4,roundToInt = T)
# 
# plotChangeMap(sc3, out3, "OLIG1") + plotChangeMap(sc4, out4, "OLIG1")
# plotChangeMap(sc3, out3, "SOX10") + plotChangeMap(sc4, out4, "SOX10")
# 
# remove(list = c("sc","sc3","sc4","out3","out4"))

# I eventually decided to use the autoEstCont(sc,contaminationRange = c(0.01,0.5)) only for this sample
lapply(id_sample2,function(x){
  # track the progress
  print(x)
  # define the location of the output of cellranger
  file <- paste0("../../data/cellranger7_out/",x)
  # Load data and estimate soup profile
  sc <- load10X(file)
  # Estimate rho
  sc <- autoEstCont(sc,contaminationRange = c(0.01,0.5))
  # Clean the data
  out <- adjustCounts(sc,roundToInt = T)
  # save the data
  DropletUtils:::write10xCounts(paste0("../../data/SoupX_default_cellranger7/",x), out)
})

# test 2 failed sample ----------------------------------------------------
# for the failed experiment I think the issue is the fact that cellranger is estimating a huge number of cells compared to the actual real number. Therefore I am re-runnign cellranger with expected (one test) and forced (another test) number of cells. the idea is to see if the issue persist. If so it would probably be better to remove that sample from the analysis.
# else I would need to run the analysis without SoupX

id_sample2 <- "W8_CSF_tolebrutinib"
# lapply(id_sample2,function(x){
#   # track the progress
#   print(x)
#   # define the location of the output of cellranger
#   file <- paste0("../../data/cellranger7_out/",x)
#   # Load data and estimate soup profile
#   sc <- load10X(file)
#   # Estimate rho
#   sc <- autoEstCont(sc,contaminationRange = c(0,0.2))
#   # Clean the data
#   out <- adjustCounts(sc,roundToInt = T)
#   # save the data
#   # DropletUtils:::write10xCounts(paste0("../../data/SoupX_default_cellranger7/",x), out)
# })
